requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
ruamel.yaml==0.18.5
aiofiles==23.2.1
jinja2==3.1.2
httpx==0.25.2
apscheduler==3.10.4
cryptography==41.0.8
python-multipart==0.0.6